+++ CHANGELOG.md
@@ -1,10 +1,19 @@
-# 0.X.X
+# Changelog
 
-## 0.2.1
+## 0.X.X
+
+### 0.3.0
+
+-   Output code is now formatted with prettier
+-   Added sidebar (`<docs-sidebar>`) (Need)
+-   All files (that aren't markdown) in `<in>` will copy to `<out>`
+-   Empty title now will be replaced by the name of the file.
+
+### 0.2.1
 
 -   Fixed a bug.
 
-## 0.2.0
+### 0.2.0
 
 -   Now using `*/index.html` instead of `*.html`
 -   Now using `<docs-content>` instead of `<!-- {content} -->`
@@ -12,6 +21,6 @@
 -   Added watch command
 -   Fixed some bugs
 
-## 0.1.0
+### 0.1.0
 
 -   Added basic things.